requirements.txt
Django==4.2.7
django-grappelli==3.0.8
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
django-environ==0.11.2
psycopg==3.1.12